feat(system): 重构字典功能并添加用户相关新功能
- 重构字典功能，新增字典项相关接口和表单
- 添加用户个人中心相关功能，包括修改密码、绑定手机和邮箱等
- 新增公共的键值对类 KeyValue
- 更新 MenuForm 类，增加路由参数等字段
- 优化 DeptForm 和 RoleForm 类的字段和注解
- 添加邮件发送功能，包括简单邮件和带附件邮件

+++ pilates-system/system-api-boot/src/main/java/com/jrmall/pilates/system/api/RoleApi.java
@@ -32,7 +32,7 @@ public interface RoleApi {
      *
      * @return {@link List<Option>} – 角色下拉列表
      */
-    List<Option<Integer>> listRoleOptions();
+    List<Option<Long>> listRoleOptions();
 
     /**
      * 保存角色
